+++ tests/Feature/Endpoints/V0/HealthCheckEndpointTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\V0\Endpoints;
+namespace Tests\Feature\Endpoints\V0;
 
 use Tests\TestCase;
 
@@ -9,7 +9,7 @@ class HealthCheckEndpointTest extends TestCase
     /** @test */
     public function it_can_check_the_server_status()
     {
-        $response = $this->get('api/v0/');
+        $response = $this->get('/v0');
 
         $response->assertStatus(200);
         $response->assertJson([
